[PAD-142] Improve MonteCarlo algorithm
Test MonteCarlo algorithm on 10 attributes.

Use a better cost model, so that the NDV of an attribute or N independent attributes smoothly approaches the size of the fact table.

Increase the number of probing queries in the Monte Carlo algorithm from 10 to 100.

Widen return type of Schema.getAttributes().

Use an ArrayDeque rather than a LinkedList; better memory usage.